requirement.txt
Django==1.11.17
djangorestframework==3.9.4
django-rest-framework==0.1.0
django-webpack-loader==0.2.4
requests==2.20.0
django-filter==2.0
pymysql
django_cors_headers==3.2.1
gradio==3.21.0
uwsgi
translate
jieba